Implement authorization using Policy

+++ app/Http/Controllers/QuestionsController.php
@@ -8,6 +8,10 @@ use App\Http\Requests\AskQuestionRequest; //importado
 
 class QuestionsController extends Controller
 {
+
+    public function __construct() {
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -71,6 +75,7 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)  //edit($id)
     {
+        $this->authorize("update", $question);
         return view("questions.edit", compact('question'));
         //$question = Question::findOrFail($id);
     }
@@ -84,6 +89,8 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        $this->authorize("update", $question); 
+
         $question->update($request->only('title', 'body'));
 
         return redirect('/questions')->with('success', "Your question has been updated.");
@@ -97,6 +104,8 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        $this->authorize("delete", $question);
+        
         $question->delete();
 
         return redirect('/questions')->with('success', "You question has been deleted.");
